+++ src/components/shared/types/swap.ts
@@ -11,3 +11,24 @@ export type SwapRouteSplitRequest = {
   from?: string; // If user is connected send this [ Connected Wallet ]
   to?: string; // If user is connected send this [ Connected Wallet ]
 };
+
+export interface SwapResponse {
+  assetIn: string;
+  assetOut: string;
+  priceImpact: {
+    numerator: string;
+    denominator: string;
+  };
+  trade: {
+    amountInMax: string;
+    amountOut: string;
+    distribution: {
+      protocol_id: string;
+      path: string[];
+      parts: number;
+      is_exact_in: boolean;
+    }[];
+    expectedAmountIn: string;
+  };
+  tradeType: "EXACT_IN" | "EXACT_OUT";
+}
